refactor(campfire): remove helpers re-export
