--- apps/campfire/src/directives/helpers.ts
@@ -1,20 +1,0 @@
-export {
-  isRange,
-  clamp,
-  convertRanges,
-  parseNumericValue,
-  parseRange,
-  getRandomInt,
-  getRandomItem,
-  getLabel,
-  stripLabel,
-  removeNode,
-  ensureKey,
-  extractAttributes,
-  type RangeValue,
-  type DirectiveNode,
-  type AttributeSpec,
-  type AttributeSchema,
-  type ExtractedAttrs,
-  type ExtractResult
-} from '@/packages/remark-campfire/helpers'
